JUDGE-xform.md merges tourney/xform-a: the DC plane re-gridded onto the
transform, n x n intra predictors, and its inverse shift chain, which tracks
the true per-size gain.  xform-b's INTRA_DIR exclusivity and its shift chain
are not taken; its fuzz seeds, its plane-cap and weighting-matrix tests, and
its lane-balance statement are.

**The transform family.**  detail-a's 4x4 and xform-a's 16x16 and 32x32 are
now the same family, `fdct_block(n)` / `idct_block(n)` over {4, 8, 16, 32},
declared once.  Two families would have been two dequantiser scales and the
second one would have been wrong SILENTLY: a factor of two shifts the
effective QP by 6 and leaves every rate and PSNR number plausible
(MERGE-PLAN 4.4).  So the decision carries `ref.transform_gain`, a new ctest
that measures every size against a floating-point orthonormal DCT-II to
within 0.1 %.

Writing that test corrected the statement of the invariant.  The 2^20 figure
is the gain of the INTERNAL fixed-point graph -- the 1D matrices are the
orthonormal ones scaled by 2^10 -- and the shift chain removes it again; the
graph grows by sqrt(2) per doubling, so the internal 2D gains are 2^20, 2^20,
2^21, 2^22 and the chains differ accordingly.  What every size must hit is
that the QUANTISER sees orthonormal coefficients at unit scale, which is the
property one qstep table depends on and which the test now measures
directly.  The test failed on its first run against my own wrong constant,
which is the point of measuring against floating point rather than against
the codec.

**split4x4 x xform_size.**  Kept as separate fields (MERGE-PLAN 4.4): one is
per tile and chooses the transform, the other is per 8x8 block and subdivides
one.  The price is the rule of SYNTAX 4.1 -- the split is meaningful only at
`xform_size == 8`, and a stream setting both is BITSTREAM -- pinned by the
new reject `r39` and recorded as Appendix A 74.

Two real defects found while merging, both silent-by-construction:

* **`NXVC_TOOL_XFORM_LARGE` came in still on bit 24**, colliding with
  `INTRA_CFL`.  The `#define` block merged cleanly, the tree compiled, and
  the encoder emitted streams in which the two tools were the same bit.  This
  is MERGE-PLAN 4.6 exactly, and what caught it was the vector generator
  refusing to decode `v68`, not the compiler and not the test suite.
* **The encoder's per-tile size search could emit the illegal combination.**
  It builds candidates by copying `tp` and overwriting `xform_size`, so a
  `split4 = 0` applied once where tile parameters are first chosen was
  silently undone for every candidate -- ASan caught it reading `kScan4Split`
  past its end at a 16x16 block.  The derivation now lives in
  `TileCoder::setup()`, which runs on every candidate and on the final tile,
  so the two fields cannot get out of step.

Composition, in one place each: `UnitCtx::level()` applies the split's band
mapping and then the transform's scan-group shift before a context is chosen,
so the detail, entropy and transform packages meet there and nowhere else;
`residual_block()` takes the size from the plane and the split from the
block; CfL's fit reads the block's own 2n neighbours, so it follows the
transform size the way every other predictor does.

SYNTAX: 6.7 is the transform size, 6.8 the 4x4 split, and every one of
detail-a's cross-references to its old 6.7 was repointed by hand.  Appendix A
gains 68-74.  6.7 now states the rANS lane-balance cost of the single-large-
unit mapping in xform-b's terms -- nine units over eight lanes on a 4:2:0
32x32 tile -- because that cost is real, it is the one criterion xform-a
loses, and silence about it is what let it through.

Checks:

* tools-off byte-identity against e4e85af: PASS, 4:4:4 and 4:2:0, QP 16/32.
* ref.* and fuzz.* 18/18 green under asan-ubsan, including the new
  ref.transform_gain.
* all 67 existing conformance vectors byte-identical; v68-v73 and r36-r39
  added; r09 moved to word1 bit 31, the reserved bit that is left.
* gain, main harness, 12 frames, vr-mixed-1024-v2, vs x264-intra:

      intra 4:4:4   base +114.98 %   xform auto +85.81 %   delta -29.17
      intra 4:2:0   base +101.56 %   xform auto +63.48 %   delta -38.08

  JUDGE-xform.md measured -29.17 and -38.08.  Reproduced to the decimal on
  the merged encoder, BD-PSNR included (-5.690 dB and -4.634 dB).

+++ ref/src/tables.cpp
@@ -55,7 +55,7 @@ const u8 kWeight[4][64] = {
     26, 28, 29, 31, 32, 32, 32, 32,
   },
 };
-const u8 kRaster8[64] = {
+const u16 kRaster8[64] = {
      0,  1,  2,  3,  4,  5,  6,  7,
      8,  9, 10, 11, 12, 13, 14, 15,
     16, 17, 18, 19, 20, 21, 22, 23,
@@ -66,13 +66,13 @@ const u8 kRaster8[64] = {
     56, 57, 58, 59, 60, 61, 62, 63,
 };
 
-const u8 kZigzag8[64] = {
+const u16 kZigzag8[64] = {
     0,  1,  8,  16, 9,  2,  3,  10, 17, 24, 32, 25, 18, 11, 4,  5,
     12, 19, 26, 33, 40, 48, 41, 34, 27, 20, 13, 6,  7,  14, 21, 28,
     35, 42, 49, 56, 57, 50, 43, 36, 29, 22, 15, 23, 30, 37, 44, 51,
     58, 59, 52, 45, 38, 31, 39, 46, 53, 60, 61, 54, 47, 55, 62, 63,
 };
-const u8 kZigzag4[16] = {0, 1, 4, 8, 5, 2, 3, 6, 9, 12, 13, 10, 7, 11, 14, 15};
+const u16 kZigzag4[16] = {0, 1, 4, 8, 5, 2, 3, 6, 9, 12, 13, 10, 7, 11, 14, 15};
 // The split-block scan (tool bit 19): the four 4x4 sub-blocks of an 8x8 block
 // in raster order, each in 4x4 zigzag, concatenated into the same 64-value
 // coding unit an unsplit block uses.  Sub-block (sx, sy) keeps its 16
@@ -80,14 +80,14 @@ const u8 kZigzag4[16] = {0, 1, 4, 8, 5, 2, 3, 6, 9, 12, 13, 10, 7, 11, 14, 15};
 // unit's coefficient layout is a bijection of the unsplit one and nothing
 // else about the unit -- CBF, LAST, the level chain, the lane schedule --
 // changes.  docs/SYNTAX.md 9.2.
-const u8 kScan4Split[64] = {
+const u16 kScan4Split[64] = {
      0,  1,  8, 16,  9,  2,  3, 10, 17, 24, 25, 18, 11, 19, 26, 27,
      4,  5, 12, 20, 13,  6,  7, 14, 21, 28, 29, 22, 15, 23, 30, 31,
     32, 33, 40, 48, 41, 34, 35, 42, 49, 56, 57, 50, 43, 51, 58, 59,
     36, 37, 44, 52, 45, 38, 39, 46, 53, 60, 61, 54, 47, 55, 62, 63,
 };
-const u8 kZigzag2[4] = {0, 1, 2, 3};
-const u8 kZigzag1[1] = {0};
+const u16 kZigzag2[4] = {0, 1, 2, 3};
+const u16 kZigzag1[1] = {0};
 // kCflRecip[d] = round(2^15 / d); entry 0 is never read (SYNTAX.md 7.7).
 const u16 kCflRecip[256] = {
         0, 32768, 16384, 10923,  8192,  6554,  5461,  4681,
